+++ dataservices/management/commands/import_countries.py
@@ -1,15 +1,42 @@
 import tablib
 from django.core.management import BaseCommand
-from import_export import resources
+from import_export import fields, resources
 
 from dataservices.models import Country
 
 
+class CountryResource(resources.ModelResource):
+    imported_rows_pk = []
+
+    class Meta:
+
+        model = Country
+        skip_unchanged = True
+        report_skipped = True
+        is_active = fields.Field(default=True)
+        import_id_fields = ['iso2']
+        fields = ('name', 'iso1', 'iso2', 'iso3', 'region', 'is_active')
+
+    def before_import(self, dataset, using_transactions, dry_run, **kwargs):
+        self.Meta.model.objects.all().update(is_active=False)
+        return super(CountryResource, self).before_import(dataset, using_transactions, dry_run, **kwargs)
+
+    def get_or_init_instance(self, instance_loader, row):
+        row['is_active'] = True
+        return super(CountryResource, self).get_or_init_instance(instance_loader, row)
+
+
 class Command(BaseCommand):
     help = 'Import Countries data'
+    DEFAULT_FILENAME = 'dataservices/resources/countries-territories-and-regions-5.35.csv'
+
+    def add_arguments(self, parser):
+        # Positional arguments
+        parser.add_argument('filename', nargs='?', type=str)
 
     def handle(self, *args, **options):
-        with open('dataservices/resources/countries-territories-and-regions-5.35.csv', 'r', encoding='utf-8-sig') as f:
+        filename = options['filename'] if options['filename'] else self.DEFAULT_FILENAME
+        with open(filename, 'r', encoding='utf-8-sig') as f:
             data = tablib.import_set(f.read(), format='csv', headers=True)
             dataset = tablib.Dataset(
                 headers=[
@@ -25,10 +52,9 @@ class Command(BaseCommand):
             for item in data:
                 if item[2] == 'Country':
                     dataset.append((item[1], item[3], item[4], item[5], item[6]))
-            country_resource = resources.modelresource_factory(model=Country)()
-
-            # Delete existing entries
-            Country.objects.all().delete()
-            country_resource.import_data(dataset)
-
-        self.stdout.write(self.style.SUCCESS('All done, bye!'))
+            country_resource = CountryResource()
+            report = country_resource.import_data(dataset)
+            [
+                self.stdout.write(self.style.SUCCESS(f'Results:  {key} : {value}'))
+                for key, value in report.totals.items()
+            ]