+++ dataservices/tests/conftest.py
@@ -5,31 +5,41 @@ import pytest
 from dataservices import models
 
 
-@pytest.fixture
-def internet_usage_data():
-    country_data = [
-        models.InternetUsage(
-            country_code='GBR',
-            country_name='United Kingdom',
-            year=2020,
-            value=90.97,
-        ),
-        models.InternetUsage(country_code='DEU', country_name='Germany', year=2020, value=91.97),
-    ]
-    yield models.InternetUsage.objects.bulk_create(country_data)
-    models.InternetUsage.objects.all().delete()
-
-
-def countries_data():
+@pytest.fixture(autouse=True)
+def countries():
+    country_uk, _ = models.Country.objects.get_or_create(iso2='GB', iso3='GBR', name='United Kingdom', iso1=826)
+    country_in, _ = models.Country.objects.get_or_create(iso2='IN', iso3='IND', name='India', iso1=200)
+    country_ca, _ = models.Country.objects.get_or_create(iso2='CA', iso3='CAN', name='Canada', iso1=201)
+    country_cn, _ = models.Country.objects.get_or_create(iso2='CN', iso3='CHN', name='China', iso1=203)
+    country_fr, _ = models.Country.objects.get_or_create(iso1='250', iso2='FR', iso3='FRA', name='France')
+    country_nl, _ = models.Country.objects.get_or_create(iso1='528', iso2='NL', iso3='NLD', name='Netherlands')
+    country_de, _ = models.Country.objects.get_or_create(iso2='DE', iso3='DEU', name='Germany')
     return {
-        'FR': models.Country.objects.create(iso1='1', iso2='FR', iso3='FRA', name='France'),
-        'NL': models.Country.objects.create(iso1='2', iso2='NL', iso3='NLD', name='Netherlands'),
+        'IN': country_in,
+        'GB': country_uk,
+        'CA': country_ca,
+        'CN': country_cn,
+        'FR': country_fr,
+        'NL': country_nl,
+        'DE': country_de,
     }
 
 
 @pytest.fixture
-def comtrade_report_data():
-    countries = countries_data()
+def internet_usage_data(countries):
+    country_data = [
+        models.InternetUsage(
+            country=countries['GB'],
+            year=2020,
+            value=90.97,
+        ),
+        models.InternetUsage(country=countries['DE'], year=2020, value=91.97),
+    ]
+    yield models.InternetUsage.objects.bulk_create(country_data)
+
+
+@pytest.fixture
+def comtrade_report_data(countries):
     models.ComtradeReport.objects.create(
         country=countries['FR'],
         country_iso3='FRA',
@@ -46,24 +56,16 @@ def comtrade_report_data():
     )
     yield
     models.ComtradeReport.objects.all().delete()
-    models.Country.objects.all().delete()
 
 
 @pytest.fixture
-def multi_country_data():
-    countries = countries_data()
+def multi_country_data(countries):
     models.EaseOfDoingBusiness.objects.create(
         country=countries['FR'],
-        country_name='France',
-        country_code='FR',
-        year_2019=12,
+        year=2019,
+        value=12,
     )
-    models.EaseOfDoingBusiness.objects.create(
-        country=countries['NL'],
-        country_code='NL',
-        country_name='Netherlands',
-        year_2019=13,
-    )
+    models.EaseOfDoingBusiness.objects.create(country=countries['NL'], year=2019, value=13)
     models.CIAFactbook.objects.create(
         country=countries['FR'],
         country_name='France',
@@ -80,12 +82,10 @@ def multi_country_data():
     yield
     models.EaseOfDoingBusiness.objects.all().delete()
     models.CIAFactbook.objects.all().delete()
-    models.Country.objects.all().delete()
 
 
 @pytest.fixture
-def age_group_data():
-    countries = countries_data()
+def age_group_data(countries):
     models.PopulationData.objects.create(
         country=countries['FR'],
         year=2019,
@@ -112,7 +112,6 @@ def age_group_data():
     )
     yield
     models.PopulationData.objects.all().delete()
-    models.Country.objects.all().delete()
 
 
 @pytest.fixture